pyproject.toml
[project]
name = "api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonpath-ng>=1.8.0",
    "openpyxl>=3.1.5",
    "pytest>=9.0.2",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
]
